Reverted jol dependency as it would crash the test

+++ internal/venice-common/src/main/java/com/linkedin/venice/common/Measurable.java
@@ -1,5 +1,5 @@
 package com.linkedin.venice.common;
 
 public interface Measurable {
-  long getSize();
+  int getSize();
 }
